requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
python-multipart==0.0.6
aiofiles==23.2.1
pathlib==1.0.1